+++ qiskit/providers/aer/backends/pulse_simulator.py
@@ -26,7 +26,7 @@ from .aerbackend import AerBackend
 from ..aerjob import AerJob
 from ..version import __version__
 from ..pulse.controllers.full_pulse_digest0 import full_digest
-from ..pulse.solver.opsolve import opsolve
+from ..pulse.pulse0.solver.opsolve import opsolve
 
 logger = logging.getLogger(__name__)
