+++ qiskit/providers/aer/pulse/pulse0/cy/utils.pyx
