+++ qiskit/providers/aer/pulse/pulse0/qobj/digest.py
