+++ qiskit/providers/aer/pulse/pulse0/qobj/operators.py
