+++ qiskit/providers/aer/pulse/pulse0/qobj/opparse.py
