+++ qiskit/providers/aer/pulse/pulse0/qutip_lite/cy/complex_math.pxi
