+++ qiskit/providers/aer/pulse/pulse0/qutip_lite/cy/sparse_structs.pxd
