+++ qiskit/providers/aer/pulse/pulse0/qutip_lite/expect.py
