+++ qiskit/providers/aer/pulse/pulse0/qutip_lite/fastsparse.py
