+++ qiskit/providers/aer/pulse/pulse0/solver/__init__.py
