+++ qiskit/providers/aer/pulse/pulse0/solver/data_config.py
