+++ qiskit/providers/aer/pulse/pulse0/solver/opsolve.py
