+++ qiskit/providers/aer/pulse/pulse0/solver/options.py
