+++ qiskit/providers/aer/pulse/pulse0/solver/rhs_utils.py
