+++ qiskit/providers/aer/pulse/pulse0/solver/settings.py
