+++ qiskit/providers/aer/pulse/pulse0/solver/unitary.py
