+++ test/terra/openpulse/test_duffing_model_generators.py
@@ -19,7 +19,7 @@ from test.terra.common import QiskitAerTestCase
 from qiskit.providers.aer.pulse.system_models.pulse_system_model import PulseSystemModel
 from qiskit.providers.aer.pulse.system_models.hamiltonian_model import HamiltonianModel
 from qiskit.providers.aer.pulse.system_models import duffing_model_generators as model_gen
-from qiskit.providers.aer.pulse.qobj.op_qobj import get_oper
+from qiskit.providers.aer.pulse.pulse0.qobj.op_qobj import get_oper
 
 class TestDuffingModelGenerators(QiskitAerTestCase):
     """Tests for functions in duffing_model_generators.py"""
